Add the functions
