+++ package.json
@@ -5,12 +5,14 @@
   "main": "index.js",
   "scripts": {
     "build": "webpack",
-    "start": "webpack serve --open"
+    "start": "webpack serve --open",
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "jsdom": "^21.1.1",
     "lodash": "^4.17.21",
     "web": "^0.0.2",
     "webpack": "^5.76.3",
@@ -24,6 +26,8 @@
     "eslint-plugin-import": "^2.27.5",
     "hint": "^7.1.6",
     "html-webpack-plugin": "^5.5.0",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
     "style-loader": "^3.3.2",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
